api/requirements.txt
flask
slack-sdk
slackeventsapi
python-dotenv
psycopg2-binary
gunicorn
